page_table: remove tracker form `PageTableInner`
For Sv48 page tables that will provide hypervisor-virtual addressing, no
interaction with page tracking is possible or desired. So break the
dependency that all page tables reflect page ownership when that's only
appropriate for moving pages to/from VMs.

This moves the `Drop` implementation logic up a layer to
`PlatformPageTable` as the page tracker releases are specific to VM page
assignements and not wanted when dropping an Sv48.