Silence unused error warning
rust has decided that only using an item for its debug output is worthy
of a warning.

this is intentional: https://github.com/rust-lang/rust/issues/123068

accept that and mark the error enums as allowing dead code to shut it
up. this is unfortunate as actual dead code warnings would be nice. If
that's a problem, we can add manual display implementations later.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>